// src/student/student.dto.ts
export class FindStudentResponseDto {
  name: string;
  teacher: string;
  id: string;
}
export class StudentResponseDto {
  name: string;
  teacher: string;
  id: string;
}
export class CreateStudentDto {
  name: string;
  teacher: string;
}
export class UpdateStudentDto {
  name: string;
  teacher: string;
}